♻️ refactor nav bar
